Rename card component
